// src/public-api/model/loginDto.ts
/**
 * MISCore Shop API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface LoginDto { 
    email: string;
    password: string;
    rememberMe?: boolean;
}