// src/public-api/model/registerDto.ts
/**
 * MISCore Shop API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RegisterDto { 
    firstName: string;
    lastName: string;
    email: string;
    password: string;
    confirmPassword?: string | null;
}