// src/public-api/model/userInfoDto.ts
/**
 * MISCore Shop API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserInfoDto { 
    id?: string | null;
    email?: string | null;
    userName?: string | null;
    firstName?: string | null;
    lastName?: string | null;
    profilePictureUrl?: string | null;
    dateOfBirth?: string | null;
    gender?: string | null;
    isActive?: boolean;
    lastLoginAt?: string | null;
    createdAt?: string;
    roles?: Array<string> | null;
}